host/topo: Fix indexing bug causing wrong GPU->NIC assignments
When a NIC that has more than 1 GPU is attempting to be rebalanced, the
IF check ((*pairs_iter).pe_idx != nic_id) confirms that the distance
between NIC/GPU is for the GPU that is trying to be rebalanced.
This indexing bug causes GPU-NIC rebalancing to happen based on
distance of the wrong GPU->NIC pair. If there are 8 PE's per node,
and 2 NICS per PE, and PE 1 is attempting to be rebalanced, this
bug would cause the distance between PE 0 -> NIC X to be used
instead of PE 1 -> NIC X.

                Seth Zegelstein <szegel@amazon.com>

+++ src/modules/transport/libfabric/libfabric.cpp
@@ -846,13 +846,13 @@ int nvshmemt_put_signal_unordered(struct nvshmem_transport *tcurr, int pe, rma_v
 
     assert(write_remote.size() == write_local.size() &&
            write_local.size() == write_bytes_desc.size());
-    for (int i = 0; i < write_remote.size(); i++) {
+    for (size_t i = 0; i < write_remote.size(); i++) {
         status =
             nvshmemt_libfabric_rma_impl(tcurr, pe, write_verb, &write_remote[i], &write_local[i],
                                         write_bytes_desc[i], is_proxy, &sequence_count);
         if (unlikely(status)) {
             NVSHMEMI_ERROR_PRINT(
-                "Error in nvshmemt_put_signal_unordered, could not submit write #%d\n", i);
+                "Error in nvshmemt_put_signal_unordered, could not submit write #%lu\n", i);
             goto out;
         }
     }
